Move buildInsertSqlQuery to Generator
